JOB PORTAL MANAGEMENT SYSTEM
The name of the project is Online Job portal

The website provides both the jobseekeers and the recruiters his PIN number which is unique.

The main stakeholders of this system are:

1. Recruiting manager.

2. Vacancies

3. Job Seekers.

4. Administrator

This system enables the Recruiting manager to login to the system and create a vacancy and post it on the web. The Recruiting manager can associate jobseekers with a vacancy and Schedule the interview. He is able to search on jobseeker number and Vacancy number.